+++ components/GiveComponents/index.js
@@ -1,0 +1,6 @@
+import GiveHero from './Hero/Hero';
+import WaysToGive from './WaysToGive';
+import WistiaCarousel from './WistiaCarousel';
+import WhatTheBibleSays from './WhatTheBibleSays';
+
+export { GiveHero, WaysToGive, WistiaCarousel, WhatTheBibleSays };
